feat: - 베리 획득 화면 수정       - 카드 즐겨찾기 수정       - 검색 아이콘 수정       - slectionUI 파일 기능별 여러 파일로 리팩토링       - 메뉴 바텀바 ui
