--- app/src/main/res/drawable/card_favorite_new.xml
@@ -1,15 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="29dp"
-    android:height="29dp"
-    android:viewportWidth="29"
-    android:viewportHeight="29">
-  <path
-      android:pathData="M14.5,14.5m-14.5,0a14.5,14.5 0,1 1,29 0a14.5,14.5 0,1 1,-29 0"
-      android:fillColor="#ffffff"
-      android:fillAlpha="0.9"/>
-  <path
-      android:strokeWidth="1"
-      android:pathData="M14.239,6.342C14.479,5.605 15.521,5.605 15.761,6.342L17.29,11.048C17.397,11.378 17.704,11.601 18.051,11.601H23C23.774,11.601 24.097,12.592 23.47,13.048L19.466,15.957C19.186,16.16 19.069,16.521 19.176,16.851L20.705,21.558C20.944,22.295 20.101,22.907 19.474,22.452L15.47,19.543C15.19,19.34 14.81,19.34 14.53,19.543L10.526,22.452C9.899,22.907 9.056,22.295 9.295,21.558L10.824,16.851C10.932,16.521 10.814,16.16 10.534,15.957L6.53,13.048C5.903,12.592 6.226,11.601 7.001,11.601H11.949C12.296,11.601 12.603,11.378 12.71,11.048L14.239,6.342Z"
-      android:fillColor="#5955FF"
-      android:strokeColor="#5955FF"/>
-</vector>
