+++ app/src/main/res/drawable/ic_board_addsection.xml
@@ -1,0 +1,16 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M19.5,12.415C19.5,10.612 19.5,9.71 18.961,9.134C18.93,9.101 18.899,9.07 18.866,9.039C18.29,8.5 17.388,8.5 15.585,8.5H14.157C13.339,8.5 12.931,8.5 12.563,8.348C12.196,8.196 11.906,7.906 11.328,7.328L10.672,6.672C10.094,6.094 9.804,5.804 9.437,5.652C9.069,5.5 8.661,5.5 7.843,5.5H7.5C5.614,5.5 4.672,5.5 4.086,6.086C3.5,6.672 3.5,7.614 3.5,9.5V14.5C3.5,16.386 3.5,17.328 4.086,17.914C4.672,18.5 5.614,18.5 7.5,18.5H14"
+      android:strokeWidth="2"
+      android:fillColor="#00000000"
+      android:strokeColor="#222222"/>
+  <path
+      android:pathData="M16,17.915H19.5M19.5,17.915H23M19.5,17.915V21.415M19.5,17.915V14.415"
+      android:strokeWidth="2"
+      android:fillColor="#00000000"
+      android:strokeColor="#222222"/>
+</vector>
