+++ app/src/main/res/drawable/ic_board_memerset.xml
@@ -1,0 +1,22 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="22dp"
+    android:height="22dp"
+    android:viewportWidth="22"
+    android:viewportHeight="22">
+  <path
+      android:strokeWidth="1"
+      android:pathData="M9.625,4.125C7.853,4.125 6.417,5.561 6.417,7.333C6.417,9.105 7.853,10.541 9.625,10.541C11.397,10.541 12.833,9.105 12.833,7.333C12.833,5.561 11.397,4.125 9.625,4.125ZM5.5,7.333C5.5,5.055 7.347,3.208 9.625,3.208C11.903,3.208 13.75,5.055 13.75,7.333C13.75,9.611 11.903,11.458 9.625,11.458C7.347,11.458 5.5,9.611 5.5,7.333Z"
+      android:fillColor="#222222"
+      android:fillType="evenOdd"
+      android:strokeColor="#222222"/>
+  <path
+      android:strokeWidth="1"
+      android:pathData="M4.116,12.824C5.577,11.363 7.559,10.542 9.626,10.542C10.56,10.542 11.477,10.71 12.337,11.029L12.018,11.888C11.26,11.607 10.45,11.459 9.626,11.459C7.802,11.459 6.054,12.183 4.764,13.472C3.583,14.653 2.877,16.219 2.766,17.875L11.001,17.875V18.792L1.834,18.792L1.834,18.334C1.834,16.267 2.655,14.285 4.116,12.824Z"
+      android:fillColor="#222222"
+      android:fillType="evenOdd"
+      android:strokeColor="#222222"/>
+  <path
+      android:pathData="M16.762,11.791H15.199L14.89,12.979C14.559,13.097 14.256,13.267 13.989,13.479L12.762,13.142L11.981,14.441L12.898,15.292C12.87,15.454 12.855,15.621 12.855,15.791C12.855,15.961 12.87,16.128 12.898,16.29L11.981,17.142L12.762,18.441L13.989,18.104C14.256,18.315 14.559,18.485 14.89,18.603L15.199,19.791H16.762L17.071,18.603C17.401,18.485 17.705,18.315 17.972,18.104L19.199,18.441L19.98,17.141L19.063,16.29C19.091,16.128 19.106,15.961 19.106,15.791C19.106,15.621 19.091,15.454 19.063,15.292L19.98,14.441L19.199,13.142L17.972,13.479C17.705,13.267 17.401,13.097 17.071,12.979L16.762,11.791ZM15.98,16.791C16.556,16.791 17.022,16.343 17.022,15.791C17.022,15.239 16.556,14.791 15.98,14.791C15.405,14.791 14.939,15.239 14.939,15.791C14.939,16.343 15.405,16.791 15.98,16.791Z"
+      android:fillColor="#222222"
+      android:fillType="evenOdd"/>
+</vector>
