+++ app/src/main/res/drawable/ic_board_rename.xml
@@ -1,0 +1,16 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M15.188,5.424C15.612,5.469 15.95,5.664 16.219,5.869C16.503,6.086 16.808,6.394 17.121,6.707L17.293,6.879C17.606,7.192 17.914,7.497 18.131,7.781C18.365,8.089 18.586,8.486 18.586,9C18.586,9.514 18.365,9.911 18.131,10.219C17.914,10.503 17.606,10.808 17.293,11.121L10.099,18.315C9.942,18.472 9.738,18.689 9.475,18.838C9.211,18.987 8.921,19.051 8.705,19.104L6.097,19.756L6.095,19.757L6.051,19.768C5.903,19.805 5.682,19.863 5.487,19.882C5.281,19.902 4.829,19.909 4.46,19.54C4.091,19.171 4.098,18.719 4.118,18.513C4.137,18.318 4.195,18.097 4.232,17.949L4.896,15.295C4.949,15.079 5.013,14.789 5.162,14.525C5.311,14.262 5.528,14.058 5.685,13.901L12.879,6.707C13.192,6.394 13.497,6.086 13.781,5.869C14.089,5.635 14.486,5.414 15,5.414L15.188,5.424Z"
+      android:strokeWidth="2"
+      android:fillColor="#00000000"
+      android:strokeColor="#222222"/>
+  <path
+      android:pathData="M12.5,7.5L16.5,11.5"
+      android:strokeWidth="1.5"
+      android:fillColor="#00000000"
+      android:strokeColor="#222222"/>
+</vector>
