+++ app/src/main/res/drawable/ic_board_selectfile.xml
@@ -1,0 +1,24 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M5.5,5.9C5.5,5.06 5.5,4.64 5.663,4.319C5.807,4.037 6.037,3.807 6.319,3.663C6.64,3.5 7.06,3.5 7.9,3.5H12.506C12.873,3.5 13.056,3.5 13.229,3.541C13.382,3.578 13.528,3.639 13.662,3.721C13.814,3.814 13.943,3.944 14.203,4.203L17.797,7.797C18.056,8.056 18.186,8.186 18.279,8.338C18.361,8.472 18.422,8.618 18.459,8.771C18.5,8.944 18.5,9.127 18.5,9.494V18.1C18.5,18.94 18.5,19.36 18.337,19.681C18.193,19.963 17.963,20.193 17.681,20.337C17.36,20.5 16.94,20.5 16.1,20.5H7.9C7.06,20.5 6.64,20.5 6.319,20.337C6.037,20.193 5.807,19.963 5.663,19.681C5.5,19.36 5.5,18.94 5.5,18.1V5.9Z"
+      android:strokeWidth="2"
+      android:fillColor="#00000000"
+      android:strokeColor="#222222"/>
+  <path
+      android:pathData="M12.5,3.5V7.1C12.5,7.94 12.5,8.36 12.663,8.681C12.807,8.963 13.037,9.193 13.319,9.337C13.64,9.5 14.06,9.5 14.9,9.5H18.5"
+      android:strokeWidth="2"
+      android:fillColor="#00000000"
+      android:strokeColor="#222222"/>
+  <path
+      android:pathData="M4,14h6v8h-6z"
+      android:fillColor="#ffffff"/>
+  <path
+      android:pathData="M4,16.5L6.5,19L11.5,14"
+      android:strokeWidth="2"
+      android:fillColor="#00000000"
+      android:strokeColor="#222222"/>
+</vector>
