+++ app/src/main/res/drawable/ic_card_favorite.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M6,0L18,0A6,6 0,0 1,24 6L24,18A6,6 0,0 1,18 24L6,24A6,6 0,0 1,0 18L0,6A6,6 0,0 1,6 0z"
+      android:fillColor="#ffffff"
+      android:fillAlpha="0.8"/>
+  <path
+      android:pathData="M11.239,4.342C11.479,3.605 12.521,3.605 12.761,4.342L14.066,8.357C14.173,8.687 14.48,8.91 14.826,8.91H19.048C19.823,8.91 20.146,9.902 19.519,10.357L16.103,12.839C15.823,13.042 15.705,13.403 15.812,13.733L17.117,17.749C17.357,18.486 16.513,19.098 15.886,18.643L12.47,16.161C12.19,15.958 11.81,15.958 11.53,16.161L8.114,18.643C7.487,19.098 6.644,18.486 6.883,17.749L8.188,13.733C8.295,13.403 8.177,13.042 7.897,12.839L4.481,10.357C3.854,9.902 4.177,8.91 4.952,8.91H9.174C9.52,8.91 9.827,8.687 9.934,8.357L11.239,4.342Z"
+      android:fillColor="#5955FF"/>
+</vector>
